feature:迁移develop分支  rpm和repository模块 代码 #208

+++ src/backend/rpm/biz-rpm/src/main/kotlin/com/tencent/bkrepo/rpm/util/xStream/repomd/RepoData.kt
@@ -43,6 +43,6 @@ data class RepoData(
     @XStreamAlias("open-checksum")
     val openChecksum: RpmChecksum,
     @XStreamAlias("open-size")
-    val openSize: Int,
+    val openSize: Long,
     val revision: String = ""
 ) : RepoIndex()
